requirements.txt
textual>=0.50.0
gitpython>=3.1.40
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.7.0
